Iter5 refactoring commit 2
