+++ internal/metrics_test.go
@@ -1,4 +1,4 @@
-package main
+package internal
 
 import (
 	"github.com/stretchr/testify/assert"
